requirements.txt
streamlit
tensorflow
tensorflow-hub
pillow
numpy
io